refactor router to seperate out test pages for long term development
